tests for ENA reader and data source
