+++ src/main/java/uk/ac/ebi/eva/contigalias/dus/ENAAssemblyReportReader.java
@@ -20,6 +20,7 @@ import uk.ac.ebi.eva.contigalias.entities.AssemblyEntity;
 import uk.ac.ebi.eva.contigalias.entities.ChromosomeEntity;
 import uk.ac.ebi.eva.contigalias.entities.ScaffoldEntity;
 
+import java.io.IOException;
 import java.io.InputStreamReader;
 import java.util.LinkedList;
 import java.util.List;
@@ -28,6 +29,33 @@ public class ENAAssemblyReportReader extends AssemblyReportReader {
 
     public ENAAssemblyReportReader(InputStreamReader inputStreamReader, boolean isScaffoldsEnabled) {
         super(inputStreamReader, isScaffoldsEnabled);
+    }
+
+    protected void parseReport() throws IOException, NullPointerException {
+        if (reader == null) {
+            throw new NullPointerException("Cannot use AssemblyReportReader without having a valid InputStreamReader.");
+        }
+        String line = reader.readLine();
+        while (line != null) {
+            if (line.startsWith("accession")) {
+                if (assemblyEntity == null) {
+                    assemblyEntity = new AssemblyEntity();
+                }
+                parseAssemblyData(line);
+            } else if (!line.startsWith("accession")) {
+                String[] columns = line.split("\t", -1);
+                if (columns.length >= 6) {
+                    if (columns[5].equals("Chromosome") && columns[3].equals("assembled-molecule")) {
+                        parseChromosomeLine(columns);
+                    } else if (isScaffoldsEnabled) {
+                        parseScaffoldLine(columns);
+                    }
+                }
+            }
+            line = reader.readLine();
+        }
+        reportParsed = true;
+        reader.close();
     }
 
     // Not present in ENA assembly reports
